feat(ui): pace chart and today's date in the Budget header, tidier breakdown table, clearer hidden transactions, Planner tweaks
- Budget: interactive pace chart sits beside the figures on wide screens (phones keep the Pace section)
- Budget: shows today's date and weekday next to the month
- Budget and History share one breakdown table: compact columns, a wide Used bar, amber for nearly-spent again
- Transactions: hidden rows show date, bank and category, with Restore all and close
- Planner: 6 months ahead, full words (Bank holidays, average, month names), Flex total aligned right

+++ src/components/planner/PlannerCalendar.tsx
@@ -140,7 +140,7 @@ export default function PlannerCalendar({ month, daysOff, onToggleDay, onStats }
             <li key={date} className="flex items-center gap-2">
               <span className="h-1.5 w-1.5 rounded-full bg-[#b39767]" />
               <span className="w-12 tabular-nums text-gray-400">
-                {new Date(`${date}T00:00:00`).toLocaleDateString("en-GB", { day: "numeric", month: "short" })}
+                {new Date(`${date}T00:00:00`).toLocaleDateString("en-GB", { day: "numeric", month: "long" })}
               </span>
               <span className="text-gray-600 dark:text-gray-300">{title}</span>
             </li>
